% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSMEDR6.R
\docType{class}
\name{bSMED}
\alias{bSMED}
\title{Class providing object with methods for bSMED}
\format{\code{\link{R6Class}} object.}
\usage{
bSMED
}
\value{
Object of \code{\link{R6Class}} with methods for running a bSMED experiment.
}
\description{
Class providing object with methods for bSMED
}
\section{Fields}{

\describe{
\item{\code{X}}{Design matrix}

\item{\code{Z}}{Responses}

\item{\code{b}}{batch size}

\item{\code{nb}}{Number of batches}

\item{\code{D}}{Dimension of data}

\item{\code{Xopts}}{Available points}

\item{\code{X0}}{Initial design}

\item{\code{package}}{Which GP package to use in IGP}

\item{\code{stats}}{List of tracked stats}

\item{\code{iteration}}{Which iteration}

\item{\code{mod}}{The GP model from IGP}
}}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/CollinErickson/bSMED}
  \item{\code{new(X, Z, corr="Gauss", verbose=0, separable=T, useC=F,useGrad=T,
         parallel=T, nug.est=T, ...)}}{This method is used to create object of this class with \code{X} and \code{Z} as the data.}

  \item{\code{update(Xnew=NULL, Znew=NULL, Xall=NULL, Zall=NULL,
restarts = 5,
param_update = T, nug.update = self$nug.est)}}{This method updates the model, adding new data if given, then running optimization again.}
  }
}
\examples{
a <- bSMED$new(D=2,L=1003,func=TestFunctions::gaussian1, obj="grad", n0=0,b=3, nb=5, X0=lhs::maximinLHS(20,2), Xopts=lhs::maximinLHS(500,2))
a$run()
}
\keyword{Gaussian}
\keyword{adaptive,}
\keyword{data,}
\keyword{experiments,}
\keyword{process,}
\keyword{regression}
\keyword{sequential,}
\keyword{simulation,}

