% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSMEDR6.R
\docType{class}
\name{bSMED}
\alias{bSMED}
\title{Class providing object with methods for bSMED}
\format{\code{\link{R6Class}} object.}
\usage{
bSMED
}
\value{
Object of \code{\link{R6Class}} with methods for running a bSMED experiment.
}
\description{
Class providing object with methods for bSMED
}
\section{Fields}{

\describe{
\item{\code{X}}{Design matrix}

\item{\code{Z}}{Responses}

\item{\code{b}}{batch size to select each iteration}

\item{\code{nb}}{Number of batches, includes the first iteration taking X0}

\item{\code{D}}{Dimension of data}

\item{\code{Xopts}}{Available points}

\item{\code{X0}}{Initial design}

\item{\code{package}}{Which GP package to use in IGP. Default is "laGP" and probably is best choice.}

\item{\code{stats}}{List of tracked stats}

\item{\code{iteration}}{Which iteration}

\item{\code{mod}}{The GP model from IGP}

\item{\code{func}}{The function to maximize}

\item{\code{func_run_together}}{Should the new points to evaluate be passed to func as a matrix of all new points in row or each separately as a vector. The matrix form (TRUE) lets you have all points passed at once to your function.}

\item{\code{func_fast}}{Whether the function is fast to evaluate and thus can be used to estimate MSE and plot it. Probably TRUE for simple tests, but FALSE for real experiments}

\item{\code{obj}}{Character telling what the objective function is, e.g. max or grad}

\item{\code{obj_func}}{Function to evaluate the objective}

\item{\code{obj_alpha}}{A parameter that can be used to have model focus on uncertain areas when model is poor. Not implemented now.}

\item{\code{scale_obj}}{If the objective is not in [0,1), then this will scale it so it works with this method. Also can choose whether use_alpha. Either way the lowest value is scaled to zero. Both choices for use_alpha do similar things, not sure which is better.}

\item{\code{n0}}{How many points were in initial design, not important since X0 must be given in}

\item{\code{alpha}}{Estimated parameter. Scalar if scale_obj==FALSE as in paper. Vector with lower and upper bounds of p if scale_obj==TRUE, this is my own addition, not in paper.}

\item{\code{use_alpha}}{Whether alpha should be estimated and used. This should be TRUE if function output is in [0,1) as required in paper. Can be FALSE if usings scale_obj.}

\item{\code{alpha_regression_i_values}}{Values used to estimate alpha (iteration)}

\item{\code{alpha_regression_p_values}}{Values used to estimate alpha (max func value)}

\item{\code{gamma}}{Power used in energy}

\item{\code{gammamax}}{A maximum value for gamma, a fraction is used based on the iteration}

\item{\code{tau}}{Threshold for design region adaptation}

\item{\code{kappa}}{Parameter for design region adaptation}

\item{\code{delta}}{A small positive number to make sure the algorithm converges. The paper doesn't mention what value to use, 1e-3 seems to work but I have no clue.}

\item{\code{Xopts_removed}}{Xopts values that were removed by the design region adaptation}

\item{\code{pX}}{p values for rows of X}

\item{\code{pXopts}}{p values for rows of Xopts}

\item{\code{qX}}{q values for rows of q}

\item{\code{qXopts}}{p values for rows of Xopts}

\item{\code{p}}{Function used to calculate energy based on the model.}

\item{\code{use_design_region_fix}}{I think it makes more sense to use the quantile of the average instead of the average of the quantiles. FALSE does what is in paper and is default.}

\item{\code{exchange_algorithm_restarts}}{How many restarts for exchange algorithm. I didn't see a difference with more restarts.}

\item{\code{parallel}}{Should the new values be calculated in parallel? Not for the model, for getting actual new Z values. Can cause trouble since I haven't used it much.}

\item{\code{parallel_cores}}{Number of cores used for parallel}

\item{\code{parallel_cluster}}{The object for the cluster currently running}

\item{\code{verbose}}{0 means print nothing, 1 means print somewhat important things, 2 means print a lot. Not much implemented though.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/CollinErickson/bSMED}
  \item{\code{new(D,
                  package="laGP", obj=NULL,
                  func, func_run_together=FALSE, func_fast=TRUE,
                  use_alpha=TRUE, scale_obj=NULL,
                  delta=1e-3,
                  use_design_region_fix = FALSE,
                  exchange_algorithm_restarts=1,
                  X0, Xopts, b, nb,
                  estimate.nugget=FALSE, set.nugget=1e-8,
                  parallel=FALSE, parallel_cores="detect",
                  verbose=0,
                  ...)}}{
                  This method is used to create object of this class.
                  User must give D, func, X0, Xopts, b, and nb.
                  }

  \item{\code{run(maxit=self$nb - self$iteration + 1,
  plotlastonly=F, noplot=F)}}{
  This method run the model. It runs maxit iterations, which defaults to the end of the experiment as determined by nb.
  plotlastonly only shows last plot, noplot shows no plots.}

  \item{\code{run1(plotit=TRUE)}}{
  This method runs a single iteration.
  It plots the results if plotit==TRUE.}

  \item{\code{plot1()}}{
  This method plots the the current model.}

  \item{\code{print_results(give_best_prediction=TRUE)}}{
  This prints the best design point found in the experiment.
  If give_best_prediction==TRUE, it prints where it thinks the maximum of the surface is.}
}
}

\examples{
set.seed(0)
a <- bSMED$new(D=2,func=TestFunctions::gaussian1, obj="func",
     n0=0,b=3, nb=5,
     X0=lhs::maximinLHS(20,2),
     Xopts=lhs::maximinLHS(500,2))
a$run()
set.seed(1)
a <- bSMED$new(D=2,func=TestFunctions::quad_peaks_slant,
     obj="func", n0=0,b=3, nb=5,
     X0=lhs::maximinLHS(20,2),
     Xopts=lhs::maximinLHS(500,2),
     use_alpha=TRUE, package="laGP",
     parallel=FALSE, func_fast=TRUE,
     use_design_region_fix=TRUE)
a$run()
}
\keyword{Gaussian}
\keyword{adaptive,}
\keyword{data,}
\keyword{experiments,}
\keyword{process,}
\keyword{regression}
\keyword{sequential,}
\keyword{simulation,}
